# swagger.yaml
swagger: "2.0"
info:
  title: Movies API
  version: "0.1"
  description: A simple REST API
consumes:
  - application/json
produces:
  - application/json
paths:
  /movies:
    post:
      tags: [Movies]
      operationId : views.movies.create_movie
      summary: Create a movie
      parameters:
        - name: payload
          in: body
          schema:
            $ref: '#/definitions/MoviePayload'
      responses:
        201:
          description: Movie created
          schema:
            $ref: '#/definitions/MoviePayload'
        502:
          description: Bad gatway
        500:
          description: Internal error
    get:
      tags: [Movies]
      operationId: views.movies.get_movies
      summary: Get list of Movies
      parameters:
        - name: title
          in: query
          type: string
          required: false
        - name: start
          in: query
          type: integer
          required: false
        - name: limit
          in: query
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: Return list of Movies
          schema:
            $ref: '#/definitions/MoviePayload'
        500:
          description: Internal error
...
securityDefinitions:
   api_key:
     type: apiKey
     name: X-Auth
     in: header
     x-apikeyInfoFunc: auth.apikey.check
...
definitions:
  MoviePayload:
    type: object
    required:
      - title
    properties:
      title:
        type: string
        description: Movie's title
        example: "Need for speed"